soptraloc_system/apps/alerts/views.py
from rest_framework import viewsets
from rest_framework.views import APIView
from rest_framework.response import Response
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework.filters import SearchFilter, OrderingFilter

# Placeholder views - se implementarán con los modelos correspondientes
class AlertViewSet(viewsets.ModelViewSet):
    # queryset = Alert.objects.all()
    # serializer_class = AlertSerializer
    pass

class NotificationViewSet(viewsets.ModelViewSet):
    # queryset = Notification.objects.all()
    # serializer_class = NotificationSerializer
    pass

from rest_framework.response import Response
from rest_framework.views import APIView


class ActiveAlertsView(APIView):
    """Vista para alertas activas."""
    def get(self, request):
        return Response({
            'message': 'Active alerts endpoint - to be implemented',
            'status': 'placeholder'
        })


class AlertHistoryView(APIView):
    """Vista para historial de alertas."""
    def get(self, request):
        return Response({
            'message': 'Alert history endpoint - to be implemented',
            'status': 'placeholder'
        })

class MarkAlertReadView(APIView):
    def post(self, request, alert_id):
        return Response({'message': 'Mark alert read endpoint - to be implemented'})